Move notification actions to a properly-named file
