+++ src/components/scenes/CurrencyNotificationScene.js
@@ -6,8 +6,8 @@ import { ScrollView } from 'react-native'
 import { Actions } from 'react-native-router-flux'
 import { sprintf } from 'sprintf-js'
 
+import { enableNotifications, fetchSettings } from '../../actions/NotificationActions.js'
 import s from '../../locales/strings.js'
-import { enableNotifications, fetchSettings } from '../../modules/Notifications/action'
 import { connect } from '../../types/reactRedux.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 import { showError } from '../services/AirshipInstance'
